feat: Completed chat service

+++ chat/internal/service/service.go
@@ -8,6 +8,6 @@ import (
 
 type ChatService interface {
 	Create(context.Context, *model.CreateChatRequest) (int, error)
-	Delete(context.Context, int64) error
+	Delete(context.Context, int) error
 	SendMessage(context.Context, *model.SendMessageRequest) error
 }
